requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
chainmap==1.0.3
charset-normalizer==2.0.12
colorama==0.4.4
combomethod==1.0.12
cryptography==36.0.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.3
importlib-resources==5.4.0
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kaitaistruct==0.9
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
nulltype==2.3.1
options==1.4.10
outcome==1.1.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prod==0.7
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pydivert==2.1.0
Pygments==2.11.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
selenium==4.1.3
selenium-wire==4.6.3
Send2Trash==1.8.0
six==1.12.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wcwidth==0.2.5
webdriver-manager==3.5.4
webencodings==0.5.1
widgetsnbextension==3.6.0
win32-setctime==1.1.0
wsproto==1.1.0
zipp==3.7.0
zstandard==0.17.0
